+++ packages/gamut-templates/src/index.ts
@@ -1,3 +1,0 @@
-export { default as Interstitial } from './Interstitial';
-export { default as GridForm } from './GridForm';
-export { default as SplitInterstitial } from './SplitInterstitial';