+++ packages/gamut-icons/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## 2.3.0 (2020-04-20)
+
+* feat(Icons): Add initial streamline icons (#702) ([52d529e](https://github.com/RyzacInc/client-modules/commit/52d529e)), closes [#702](https://github.com/RyzacInc/client-modules/issues/702)
+
+
+
+
+
+## <small>2.2.2 (2020-04-13)</small>
+
+* fix(Icon): Fixes type incongruence from compiled definitions  (#720) ([d09e758](https://github.com/RyzacInc/client-modules/commit/d09e758)), closes [#720](https://github.com/RyzacInc/client-modules/issues/720)
+
+
+
+
+
+## <small>2.2.1 (2020-04-06)</small>
+
+* fix(Build): Update lockfile and config file to include new template (#699) ([9060cd2](https://github.com/RyzacInc/client-modules/commit/9060cd2)), closes [#699](https://github.com/RyzacInc/client-modules/issues/699)
+
+
+
+
+
+## 2.2.0 (2020-04-06)
+
+* feat(Icons): Add named export to gamut icons and update SVGR (#697) ([9f783da](https://github.com/RyzacInc/client-modules/commit/9f783da)), closes [#697](https://github.com/RyzacInc/client-modules/issues/697)
+
+
+
+
+
 ## [2.1.5](https://github.com/RyzacInc/client-modules/compare/@codecademy/gamut-icons@2.1.4...@codecademy/gamut-icons@2.1.5) (2020-01-15)
 
 **Note:** Version bump only for package @codecademy/gamut-icons